+++ private/lib/WinRT/GUIDParser.cpp
@@ -455,8 +455,8 @@ HRESULT GUIDParser::TryParseGUIDWithNoStyle(LPCWSTR g, GUID* result)
 
 bool GUIDParser::IsWhitespace(wchar_t c)
 {
-    int length = wcslen(whitespace);
-    for (int i = 0; i < length; i++)
+    size_t length = wcslen(whitespace);
+    for (size_t i = 0; i < length; i++)
     {
         if (c == whitespace[i])
         {
